terraform-builder/cloud-function/requirements.txt
google-cloud-build==2.0.0